Create NCR2.0.sol
Updated NeosCredits smart contract to improve security and fix exploits.

This update includes fixes for various security vulnerabilities, including the use of SafeMath to prevent integer overflow and underflow. Additionally, the transfer function has been updated to only be payable when sending Ether, and the isContract function has been replaced with the built-in code is not in order to avoid potential issues.